Fixing errors
