+++ assets/src/components/apps/app/components/kubernetes/nodes/NodesSideNav.tsx
@@ -1,7 +1,7 @@
 import { Tab, TabList } from '@pluralsh/design-system'
 import { useLocation } from 'react-router-dom'
 
-import { LinkTabWrap } from '../../utils/Tabs'
+import { LinkTabWrap } from 'components/utils/Tabs'
 
 const DIRECTORY = [
   { path: '/nodes', label: 'Nodes' },
