+++ assets/src/components/utils/events.ts
@@ -1,0 +1,4 @@
+export const ignoreEvent = (e:Event) => {
+  e.stopPropagation()
+  e.preventDefault()
+}
